Refactor: added services and controllers in register
