+++ src/config/passport.config.js
@@ -3,7 +3,7 @@ import local from 'passport-local';
 import GitHubStrategy from 'passport-github2';
 import configObject from './config.js';
 import UserModel from '../dao/models/user.models.js';
-import encrypt from '../utils/encrypt.js';
+import encrypt from '../helpers/encrypt.js';
 
 const env = configObject;
 const LocalStrategy = local.Strategy;
