+++ src/helpers/encrypt.js
