+++ src/services/session.services.js
@@ -1,0 +1,24 @@
+import UserModel from '../dao/models/user.models.js';
+import encrypt from '../helpers/encrypt.js';
+
+class UserService {
+  constructor() {
+    this.UserModel = UserModel;
+  }
+
+  createUser = async (userData) => {
+    const {
+      firstname, lastname, email, age, password,
+    } = userData;
+    const hashedPassword = await encrypt.createHash(password);
+    return this.UserModel.create({
+      firstname,
+      lastname,
+      email,
+      age,
+      password: hashedPassword,
+    });
+  };
+}
+
+export default UserService;
